requirements.txt
requests 
python-dotenv 
discord.py